refactor(position): 캐싱되어 있는 값을 반환하도록 변경
