+++ src/main/java/chess/domain/piece/Empty.java
@@ -2,6 +2,7 @@ package chess.domain.piece;
 
 import chess.domain.Color;
 import chess.domain.Name;
+import chess.domain.position.Position;
 
 public class Empty extends Piece {
     public Empty() {
