+++ src/main/java/chess/domain/piece/Pieces.java
@@ -1,6 +1,7 @@
 package chess.domain.piece;
 
 import chess.domain.Color;
+import chess.domain.position.Position;
 
 import java.util.*;
