+++ src/main/java/chess/domain/piece/Queen.java
@@ -1,9 +1,10 @@
 package chess.domain.piece;
 
 import chess.domain.Color;
-import chess.domain.Cross;
-import chess.domain.Diagonal;
+import chess.domain.position.Cross;
+import chess.domain.position.Diagonal;
 import chess.domain.Name;
+import chess.domain.position.Position;
 
 import java.util.ArrayList;
 import java.util.List;
