+++ src/test/java/chess/domain/piece/BishopTest.java
@@ -1,7 +1,8 @@
 package chess.domain.piece;
 
 import chess.domain.Color;
-import chess.domain.Diagonal;
+import chess.domain.position.Diagonal;
+import chess.domain.position.Position;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
