+++ src/test/java/chess/domain/piece/EmptyTest.java
@@ -1,5 +1,6 @@
 package chess.domain.piece;
 
+import chess.domain.position.Position;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
